# Conflicts:
#	Assets/Scripts/App/Controllers/AbilitiesController.cs
#	Assets/Scripts/App/Controllers/BattleController.cs
#	Assets/Scripts/App/Controllers/RanksController.cs
#	Assets/Scripts/App/Gameplay/Abilities/CostsLessAbility.cs
#	Assets/Scripts/App/Gameplay/Player.cs
#	Assets/Scripts/App/Managers/OverlordExperienceManager.cs
#	Assets/Scripts/App/Model/OverlordData.cs
#	Assets/Scripts/App/Pages/Items/OverlordSkillSelectionTab.cs
#	Assets/Scripts/Tests/EditMode/DataTest.cs

+++ Assets/Scripts/App/Pages/MainMenuWithNavigationPage.cs
@@ -168,7 +168,7 @@ namespace Loom.ZombieBattleground
             {
                 deck = _uiManager.GetPopup<DeckSelectionPopup>().GetDefaultDeck();
             }
-            _uiManager.GetPage<GameplayPage>().CurrentDeckId = (int)deck.Id;
+            _uiManager.GetPage<GameplayPage>().CurrentDeckId = deck.Id;
             GameClient.Get<IGameplayManager>().CurrentPlayerDeck = deck;
             GameClient.Get<IMatchManager>().FindMatch();
